+++ Projectiles/Minions/Necromancer/Necromancer.cs
@@ -357,7 +357,7 @@ namespace AmuletOfManyMinions.Projectiles.Minions.Necromancer
 		public override void CheckActive()
 		{
 			base.CheckActive();
-			if(player.ownedProjectileCounts[CounterType] == 0)
+			if(player.ownedProjectileCounts[CounterType] == 0 && animationFrame > 2)
 			{
 				projectile.Kill();
 			}
